company detail page complete

+++ Betagig/Company.swift
@@ -18,7 +18,7 @@ class Company {
     var city: String = ""
     var state: String = ""
     var zip: String = ""
-    var cost: Double = 150.00
+    var cost: Int = 150
     var shortdesc: String = ""
     var longdesc: String = ""
     var image: String = ""
@@ -37,7 +37,7 @@ class Company {
         city = snapshot.value["city"] as! String
         state = snapshot.value["state"] as! String
         zip = snapshot.value["zip"] as! String
-        cost = snapshot.value["cost"] as! Double
+        cost = snapshot.value["cost"] as! Int
         shortdesc = snapshot.value["shortdesc"] as! String
         longdesc = snapshot.value["longdesc"] as! String
         image = snapshot.value["image"] as! String
@@ -64,7 +64,7 @@ class Company {
     }
     
     
-    init (name: String, gigs: [String], region: String, street: String, city: String, state: String, zip: String, cost: Double, shortdesc: String, longdesc: String, image: String, key: String = "") {
+    init (name: String, gigs: [String], region: String, street: String, city: String, state: String, zip: String, cost: Int, shortdesc: String, longdesc: String, image: String, key: String = "") {
         
         self.name = name
         self.gigs = gigs
